requirements.txt
Django==5.0
djangorestframework==3.15.2
psycopg2==2.9.10
flake8==7.1